requirements.txt
gradio
openai
python-dotenv
chromadb
tiktoken
PyPDF2
python-docx
pytest
reportlab